Template tags -- Updated code that evaluates back base state.
With the new system using the set type slug as application instance namespace the code had to be updated to use the correct variable name cat_set_type. Additionally, instead of populating the variable with a try/except blog it now just uses the get_root() method provided by treebeard to get to the slug value.