Set model -- Ensure populated date field.
Before this commit, when the user removes the date from the DateInputField, the save method didn't take the default of datetime.now() into account but interpreted the field value as null. This led to a null constraint error on saving. Fixed by checking against null in the save method. Resolves #106.