Items -- Added file deletion on item delete.
Deleting an item from a filerset now deletes the file from filer (and the file system) as well and keeps the directory structure and sets in sync. This is a shift towards a reduced functionality scope of django-filersts. Reusing files in multiple sets is still a conceptual issue, but will most probably be a) enabled by an alternate app layer or (e.g. duplicating files with the ability to update all duplicates in one turn) b) completely scraped from filersets.  Resolved #132.